+++ src/sagemaker/estimator.py
@@ -310,7 +310,8 @@ class EstimatorBase(with_metaclass(ABCMeta, object)):
         estimator.latest_training_job.wait()
         return estimator
 
-    def deploy(self, initial_instance_count, instance_type, endpoint_name=None, use_compiled_model=False, **kwargs):
+    def deploy(self, initial_instance_count, instance_type, accelerator_type=None, endpoint_name=None,
+               use_compiled_model=False, **kwargs):
         """Deploy the trained model to an Amazon SageMaker endpoint and return a ``sagemaker.RealTimePredictor`` object.
 
         More information:
@@ -320,6 +321,10 @@ class EstimatorBase(with_metaclass(ABCMeta, object)):
             initial_instance_count (int): Minimum number of EC2 instances to deploy to an endpoint for prediction.
             instance_type (str): Type of EC2 instance to deploy to an endpoint for prediction,
                 for example, 'ml.c4.xlarge'.
+            accelerator_type (str): Type of Elastic Inference accelerator to attach to an endpoint for model loading
+                and inference, for example, 'ml.eia1.medium'. If not specified, no Elastic Inference accelerator
+                will be attached to the endpoint.
+                For more information: https://docs.aws.amazon.com/sagemaker/latest/dg/ei.html
             endpoint_name (str): Name to use for creating an Amazon SageMaker endpoint. If not specified, the name of
                 the training job is used.
             use_compiled_model (bool): Flag to select whether to use compiled (optimized) model. Default: False.
@@ -345,6 +350,7 @@ class EstimatorBase(with_metaclass(ABCMeta, object)):
         return model.deploy(
             instance_type=instance_type,
             initial_instance_count=initial_instance_count,
+            accelerator_type=accelerator_type,
             endpoint_name=endpoint_name)
 
     @property