+++ src/sagemaker/fw_utils.py
@@ -39,11 +39,12 @@ EMPTY_FRAMEWORK_VERSION_ERROR = 'framework_version is required for script mode e
                                 'Please add framework_version={} to your constructor to avoid this error.'
 
 VALID_PY_VERSIONS = ['py2', 'py3']
+VALID_EIA_FRAMEWORKS = ['tensorflow', 'mxnet']
+VALID_ACCOUNTS_BY_REGION = {'us-gov-west-1': '246785580436'}
 
 
 def create_image_uri(region, framework, instance_type, framework_version, py_version=None,
-                     account='520713654638', optimized_families=None):
-
+                     account='520713654638', accelerator_type=None, optimized_families=None):
     """Return the ECR URI of an image.
 
     Args:
@@ -54,6 +55,7 @@ def create_image_uri(region, framework, instance_type, framework_version, py_ver
         py_version (str): Optional. Python version. If specified, should be one of 'py2' or 'py3'.
             If not specified, image uri will not include a python component.
         account (str): AWS account that contains the image. (default: '520713654638')
+        accelerator_type (str): SageMaker Elastic Inference accelerator type.
         optimized_families (str): Instance families for which there exist specific optimized images.
 
     Returns:
@@ -65,8 +67,7 @@ def create_image_uri(region, framework, instance_type, framework_version, py_ver
         raise ValueError('invalid py_version argument: {}'.format(py_version))
 
     # Handle Account Number for Gov Cloud
-    if region == 'us-gov-west-1':
-        account = '246785580436'
+    account = VALID_ACCOUNTS_BY_REGION.get(region, account)
 
     # Handle Local Mode
     if instance_type.startswith('local'):
@@ -90,8 +91,32 @@ def create_image_uri(region, framework, instance_type, framework_version, py_ver
         tag = "{}-{}-{}".format(framework_version, device_type, py_version)
     else:
         tag = "{}-{}".format(framework_version, device_type)
+
+    if _accelerator_type_valid_for_framework(framework=framework, accelerator_type=accelerator_type,
+                                             optimized_families=optimized_families):
+        framework += '-eia'
+
     return "{}.dkr.ecr.{}.amazonaws.com/sagemaker-{}:{}" \
         .format(account, region, framework, tag)
+
+
+def _accelerator_type_valid_for_framework(framework, accelerator_type=None, optimized_families=None):
+    if accelerator_type is None:
+        return False
+
+    if framework not in VALID_EIA_FRAMEWORKS:
+        raise ValueError('{} is not supported with Amazon Elastic Inference. Currently only '
+                         'TensorFlow and MXNet are supported for SageMaker.'.format(framework))
+
+    if optimized_families:
+        raise ValueError('Neo does not support Amazon Elastic Inference.')
+
+    if not accelerator_type.startswith('ml.eia') and not accelerator_type == 'local_sagemaker_notebook':
+        raise ValueError('{} is not a valid SageMaker Elastic Inference accelerator type. '
+                         'See: https://docs.aws.amazon.com/sagemaker/latest/dg/ei.html'
+                         .format(accelerator_type))
+
+    return True
 
 
 def validate_source_dir(script, directory):