+++ tests/unit/test_create_deploy_entities.py
@@ -27,6 +27,7 @@ FULL_CONTAINER_DEF = {'Environment': {}, 'Image': IMAGE, 'ModelDataUrl': 's3://m
 VPC_CONFIG = {'Subnets': ['subnet-foo'], 'SecurityGroups': ['sg-foo']}
 INITIAL_INSTANCE_COUNT = 1
 INSTANCE_TYPE = 'ml.c4.xlarge'
+ACCELERATOR_TYPE = 'ml.eia.medium'
 REGION = 'us-west-2'
 
 
@@ -66,7 +67,25 @@ def test_create_endpoint_config(sagemaker_session):
     expected_pvs = [{'ModelName': MODEL_NAME,
                      'InitialInstanceCount': INITIAL_INSTANCE_COUNT,
                      'InstanceType': INSTANCE_TYPE,
+                     'InitialVariantWeight': 1,
                      'VariantName': 'AllTraffic'}]
+    sagemaker_session.sagemaker_client.create_endpoint_config.assert_called_once_with(
+        EndpointConfigName=ENDPOINT_CONFIG_NAME, ProductionVariants=expected_pvs)
+
+
+def test_create_endpoint_config_with_accelerator(sagemaker_session):
+    returned_name = sagemaker_session.create_endpoint_config(name=ENDPOINT_CONFIG_NAME, model_name=MODEL_NAME,
+                                                             initial_instance_count=INITIAL_INSTANCE_COUNT,
+                                                             instance_type=INSTANCE_TYPE,
+                                                             accelerator_type=ACCELERATOR_TYPE)
+
+    assert returned_name == ENDPOINT_CONFIG_NAME
+    expected_pvs = [{'ModelName': MODEL_NAME,
+                     'InitialInstanceCount': INITIAL_INSTANCE_COUNT,
+                     'InstanceType': INSTANCE_TYPE,
+                     'InitialVariantWeight': 1,
+                     'VariantName': 'AllTraffic',
+                     'AcceleratorType': ACCELERATOR_TYPE}]
     sagemaker_session.sagemaker_client.create_endpoint_config.assert_called_once_with(
         EndpointConfigName=ENDPOINT_CONFIG_NAME, ProductionVariants=expected_pvs)